requirements.txt
# Core scientific computing
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0

# GARCH modeling
arch>=6.0.0

# Data fetching
requests>=2.28.0
yfinance>=0.2.0
fredapi>=0.5.0

# Dashboard
streamlit>=1.28.0
plotly>=5.15.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Utilities
tqdm>=4.65.0
